Advanced Programming Principles Assignment #2
The purpose of this assignment is to use a dynamic array structure to take the user input of 15 random numbers and insert them into an array. After the numbers are inserted into an array, the bubble sort sorting algorithm is used to sort the numbers from least to greatest. After we sort the numbers we use a binary search to search the array for a value retrieved from prompting the user to pick a number. If the number is not found in the array then the program outputs  " __ was not found in the array". If the user input value is found then the user receives an output that their number was found and the what index that number is found in