Create Assignment #3
(ill admit...not my best work.. still got 100%). The purpose of this assignment was to make an HTML converter in c++. To accomplish the given task we must take text from the file “input.txt” and transform it into html then insert this text into a file named “Assignment3.html”. When opened, assignment3.html will display an html webpage with a table. This table had to have a title, and two rows of text. One row of the table must be blue and the other row must be red.