Advanced Programming Principles Assignment #4
The purpose of this assignment is to create a program that takes player information including name, number, and points scored on 6 different players. After getting the information of the player, the information is then put into a table. We then must show the total number of points scored by all players and the information of the player that scored the most points.